requirements.txt
zeroconf
python-socketio
aiohttp
